Do not call tearDown() when test is skipped

+++ tests/database/mysql/introspection.php
@@ -43,6 +43,9 @@ class Database_MySQL_Introspection_Test extends PHPUnit_Framework_TestCase
 
 	public function tearDown()
 	{
+		if ( ! $this->_db)
+			return;
+
 		$this->_db->execute_command('DROP TABLE IF EXISTS '.$this->_table);
 
 		$this->_db->disconnect();
